added function to rename or remove loadsets
These functions are factored out into those parts that rename the transcript and file storage, and those that do a search and replace in command files.

The higher level function is rename_load which will rename the .T file, the associated directory, and then replace the old with new name in all K files.